build: enforce warning-clean compilation

+++ include/curve.h
@@ -52,10 +52,11 @@ Curve createTangent(double amplitude = 1.0,
 					char curveChar = getRandomLetter(), 
 					Color color = getRandomColor());
 
-Curve createPolynomial(std::size_t order = getRandomNumber(1, 5),
-					   char curveChar = getRandomLetter(),
-					   Color color = getRandomColor());
-
+Curve createPolynomial(
+	std::size_t order = static_cast<std::size_t>(getRandomNumber(1, 5)),
+	char curveChar = getRandomLetter(),
+	Color color = getRandomColor()
+);
 
 Curve createExponential(double verticalScale = 1.0,
 						double base = 2.0,
